requirements.txt
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.1
bleach==6.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.18
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipympl==0.9.3
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.14.2
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mpmath==1.3.0
narwhals==1.30.0
nbclassic==1.2.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
open3d==0.13.0
opencv-python==4.11.0.86
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
portalocker==3.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytorch3d==0.3.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
ruff==0.11.0
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.14
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
wandb==0.19.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.13
y-py==0.6.2
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.20.2
